requirements.txt
streamlit
mistralai
pillow
requests
python-dotenv
